Reference

​

Contents

​

Reference

Contents

Index

​

Index

​

SolverLogger.Logger

SolverLogger.header

SolverLogger.row

​

SolverLogger.Logger — Type

Logger(
    :key1 => (name, fmt),
    :key2 => (name, fmt),
    …;
    kwargs...
)

Creates a solver logger.

Inputs

Required:

Arbitrary number of pairs of Symbol (key) to Tuple{String,String} (name, formatting string). The keys are currently not used, but could be in the future. The names are used for [header](@ref)s. The formatting string is a C printf type of formatting user for [row](@ref)s.

Optional, by keyword:

mode :: Symbol: Define the logging operation. Valid options are

print: Prints to stdout using println.

(TODO) save: Prints to a file.

(TODO) string: Returns the string.

verbosity :: Int: Verbosity level.

0 means silent

Greater than 0 means verbose

(TODO) Different values

Output

A Logger object.

SolverLogger.header — Method

header(logger)

Print the header of a Logger.

SolverLogger.row — Method

row(logger, args...)
logger(args...)

A Logger object is callable. Calling it will print a row of the logger.

Inputs

Required

Values corresponding to the keys used to create the logger.